requirements.txt
Flask
scikit-learn
pandas
newspaper3k
lxml_html_clean
